Split of rules & rule tests
