+++ package.json
@@ -57,6 +57,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.0.3",
+    "diff": "^4.0.2",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.0",
     "eslint-plugin-import": "^2.20.1",
