+++ src/rules/BracesForIfElse.js
@@ -1,0 +1,14 @@
+// Braces for if else
+
+const { findRangeBetweenStrings } = require("../utils");
+
+const rule = {
+    range: {
+        type: "function",
+        func: (_errLine, errItem, _evaluatedVars, allLines) => {
+            return findRangeBetweenStrings(allLines, errItem, "if", "{");
+        }
+    }
+};
+
+module.exports = { rule };
