+++ src/rules/ConsecutiveBlankLines.js
@@ -1,0 +1,62 @@
+// Consecutive blank lines
+
+const { getVariable } = require("../utils");
+
+const rule = {
+    scope: "file",
+    fix: {
+        label: "Remove consecutive blank lines",
+        type: "function",
+        func: (allLines, variables) => {
+            const lineNumber = getVariable(variables, "lineNb", { mandatory: true });
+            const newFileLines = [];
+            let prevLine = "none";
+            let pos = 0;
+            let alreadySkippedLine = false;
+            for (const line of allLines) {
+                // Check if previous line is empty: if not do not add line
+                if (line.trim() === "" && prevLine.trim() === "" && alreadySkippedLine === false && pos >= lineNumber) {
+                    alreadySkippedLine = true;
+                    pos++;
+                    continue;
+                }
+                newFileLines.push(line);
+                prevLine = line;
+                pos++;
+            }
+            return newFileLines;
+        }
+    },
+    tests: [
+        {
+            sourceBefore: `
+if (a == 2){
+
+
+x = 1
+}`,
+            sourceAfter: `
+if (a == 2){
+
+    x = 1
+}`
+        },
+        {
+            sourceBefore: `
+if (a == 2){
+
+
+
+
+x = 1
+}`,
+            sourceAfter: `
+if (a == 2){
+
+    x = 1
+}`
+        }
+    ]
+};
+
+module.exports = { rule };
