+++ src/rules/SystemExit.js
@@ -1,0 +1,14 @@
+// System.exit forbidden
+
+const { getVariableRange } = require("../utils");
+
+const rule = {
+    range: {
+        type: "function",
+        func: (errLine, errItem, evaluatedVars) => {
+            return getVariableRange(errLine, evaluatedVars, "System.exit", errItem);
+        }
+    }
+};
+
+module.exports = { rule };
