+++ test/helpers/init.js
@@ -1,1 +1,6 @@
 console.log('npm run test initialized');
+// Activate debug log if we are in debug mode
+const debug = typeof v8debug === 'object' || /--debug|--inspect|--inspect-brk/.test(process.execArgv.join(' '));
+if (debug) {
+    require('debug').enable('npm-groovy-lint')
+}
